do i still have this token
